feat: validation refactored with class-validator
