+++ src/student/student.controller.ts
@@ -1,23 +1,16 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Body, Patch, Param, Delete } from '@nestjs/common';
 import { StudentService } from './student.service';
-import {UpdatetudentProfileDto} from "./dto/update-student.dto";
-
-
+import { UpdatetudentProfileDto } from './dto/update-student.dto';
 
 @Controller('student')
 export class StudentController {
   constructor(private readonly studentService: StudentService) {}
 
-  // @Post('/add-student')  // Student jest tworzony przez Admina w UserSerwice
-  // create() {
-  //   return this.studentService.();
-  // }
-
-  @Patch('update/:id')
+  @Patch(':studentId')
   updateProfile(
-      @Body()updateProfile: UpdatetudentProfileDto,
-      @Param('studentId') studentId: string,
-  ):Promise<UpdatetudentProfileDto>{
+    @Body() updateProfile: UpdatetudentProfileDto,
+    @Param('studentId') studentId: string,
+  ): Promise<UpdatetudentProfileDto> {
     return this.studentService.updateProfile(studentId, updateProfile);
   }
